refactor(headless): split Accordion into per-file parts for tree-shaking
