+++ packages/headless/src/primitives/accordion/index.ts
@@ -1,8 +1,9 @@
+export * as Accordion from './parts';
+
 export type {
   AccordionHeaderProps,
   AccordionItemProps,
   AccordionPanelProps,
   AccordionProps,
   AccordionTriggerProps,
-} from './accordion';
-export { Accordion } from './accordion';
+} from './parts';
