+++ packages/headless/src/primitives/accordion/parts.ts
@@ -1,0 +1,5 @@
+export { type AccordionProps, AccordionRoot as Root } from './accordion-root';
+export { type AccordionItemProps, AccordionItem as Item } from './accordion-item';
+export { type AccordionHeaderProps, AccordionHeader as Header } from './accordion-header';
+export { type AccordionTriggerProps, AccordionTrigger as Trigger } from './accordion-trigger';
+export { type AccordionPanelProps, AccordionPanel as Panel } from './accordion-panel';
